Extract the glue job into an own type
